fix(architecture): resolve domain-to-data/ui layer boundary violations
Move CapabilityTag.labelOf out of domain (it depended on the UI-layer
L10n type) into a new ui/utils/capability_tag_presenter.dart extension,
following the existing presenter pattern. Relocate BookkeepingRefreshCase
from shared lib/domain/use_case/ into features/settings/domain/use_case/,
matching precedent for feature-scoped use cases that depend on their own
repositories directly. Drop an unused doc-only import in
platform_repository_port.dart.
