+++ packages/client/lib/domain/capability/capability_tag.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:tentura/ui/l10n/l10n.dart';
 
 import 'capability_group.dart';
 
@@ -21,7 +20,10 @@ enum CapabilityTag {
   // Knowledge
   localKnowledge(slug: 'local_knowledge', group: CapabilityGroup.knowledge),
   legalNavigation(slug: 'legal_navigation', group: CapabilityGroup.knowledge),
-  medicalNavigation(slug: 'medical_navigation', group: CapabilityGroup.knowledge),
+  medicalNavigation(
+    slug: 'medical_navigation',
+    group: CapabilityGroup.knowledge,
+  ),
   documents(slug: 'documents', group: CapabilityGroup.knowledge),
   verification(slug: 'verification', group: CapabilityGroup.knowledge),
 
@@ -68,46 +70,6 @@ enum CapabilityTag {
   };
 
   static CapabilityTag? fromSlug(String slug) => _bySlug[slug];
-
-  String labelOf(L10n l10n) => switch (this) {
-    CapabilityTag.transport => l10n.capabilityTagTransport,
-    CapabilityTag.storage => l10n.capabilityTagStorage,
-    CapabilityTag.pickupDelivery => l10n.capabilityTagPickupDelivery,
-    CapabilityTag.tools => l10n.capabilityTagTools,
-    CapabilityTag.physicalHelp => l10n.capabilityTagPhysicalHelp,
-    CapabilityTag.calls => l10n.capabilityTagCalls,
-    CapabilityTag.translation => l10n.capabilityTagTranslation,
-    CapabilityTag.writing => l10n.capabilityTagWriting,
-    CapabilityTag.negotiation => l10n.capabilityTagNegotiation,
-    CapabilityTag.introductions => l10n.capabilityTagIntroductions,
-    CapabilityTag.localKnowledge => l10n.capabilityTagLocalKnowledge,
-    CapabilityTag.legalNavigation => l10n.capabilityTagLegalNavigation,
-    CapabilityTag.medicalNavigation => l10n.capabilityTagMedicalNavigation,
-    CapabilityTag.documents => l10n.capabilityTagDocuments,
-    CapabilityTag.verification => l10n.capabilityTagVerification,
-    CapabilityTag.pets => l10n.capabilityTagPets,
-    CapabilityTag.childcare => l10n.capabilityTagChildcare,
-    CapabilityTag.eldercare => l10n.capabilityTagEldercare,
-    CapabilityTag.emotionalSupport => l10n.capabilityTagEmotionalSupport,
-    CapabilityTag.hosting => l10n.capabilityTagHosting,
-    CapabilityTag.money => l10n.capabilityTagMoney,
-    CapabilityTag.orders => l10n.capabilityTagOrders,
-    CapabilityTag.gig => l10n.capabilityTagGig,
-    CapabilityTag.job => l10n.capabilityTagJob,
-    CapabilityTag.food => l10n.capabilityTagFood,
-    CapabilityTag.housing => l10n.capabilityTagHousing,
-    CapabilityTag.equipment => l10n.capabilityTagEquipment,
-    CapabilityTag.workspace => l10n.capabilityTagWorkspace,
-    CapabilityTag.techHelp => l10n.capabilityTagTechHelp,
-    CapabilityTag.repair => l10n.capabilityTagRepair,
-    CapabilityTag.manualWork => l10n.capabilityTagManualWork,
-    CapabilityTag.software => l10n.capabilityTagSoftware,
-    CapabilityTag.design => l10n.capabilityTagDesign,
-    CapabilityTag.adminPaperwork => l10n.capabilityTagAdminPaperwork,
-    CapabilityTag.time => l10n.capabilityTagTime,
-    CapabilityTag.contact => l10n.capabilityTagContact,
-    CapabilityTag.other => l10n.capabilityTagOther,
-  };
 
   IconData get icon => switch (this) {
     CapabilityTag.transport => Icons.directions_car_rounded,
