+++ packages/client/lib/features/beacon/ui/widget/coordination_ui.dart
@@ -7,6 +7,7 @@ import 'package:tentura/design_system/tentura_design_system.dart';
 import 'package:tentura/domain/capability/capability_tag.dart';
 import 'package:tentura/domain/entity/coordination_response_type.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
+import 'package:tentura/ui/utils/capability_tag_presenter.dart';
 
 // --- Semantic colors (on neutral [ColorScheme.surface] / body text) ------------
 //
