+++ packages/client/lib/features/beacon_view/ui/dialog/help_offer_message_dialog.dart
@@ -9,6 +9,7 @@ import 'package:tentura/domain/capability/offer_help_suggestions.dart';
 import 'package:tentura/domain/entity/withdraw_reason.dart';
 import 'package:tentura/features/capability/ui/widget/capability_chip_set.dart';
 import 'package:tentura/features/capability/ui/widget/capability_tag_chip.dart';
+import 'package:tentura/ui/utils/capability_tag_presenter.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
 import 'package:tentura/ui/test_ids.dart';
 import 'package:tentura/ui/widget/tentura_info_hint_button.dart';
