+++ packages/client/lib/features/capability/ui/widget/capability_chip_set.dart
@@ -5,6 +5,7 @@ import 'package:tentura/domain/capability/capability_group.dart';
 import 'package:tentura/domain/capability/capability_tag.dart';
 import 'package:tentura/features/capability/ui/widget/capability_selection_count_badge.dart';
 import 'package:tentura/features/capability/ui/widget/capability_tag_chip.dart';
+import 'package:tentura/ui/utils/capability_tag_presenter.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
 import 'package:tentura/ui/widget/accordion_expansion.dart';
