+++ packages/client/lib/ui/widget/beacon_identity_tile.dart
@@ -4,6 +4,7 @@ import 'package:tentura/design_system/tentura_design_system.dart';
 import 'package:tentura/domain/entity/beacon.dart';
 import 'package:tentura/domain/entity/beacon_cover.dart';
 import 'package:tentura/domain/entity/image_entity.dart';
+import 'package:tentura/ui/utils/capability_tag_presenter.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
 
 /// Resolved request identity: cover photo, capability symbol, or neutral glyph.
