+++ packages/client/lib/ui/widget/beacon_requirements_bar.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:tentura/design_system/tentura_design_system.dart';
 import 'package:tentura/domain/capability/capability_tag.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
+import 'package:tentura/ui/utils/capability_tag_presenter.dart';
 
 /// Read-only row of capability icons for [needs] slugs (comma-split beacon field).
 ///
